fix: fieldMappingTime not work in grid form
