+++ packages/effects/plugins/src/vxe-table/use-vxe-grid.vue
@@ -244,7 +244,7 @@ async function init() {
   const autoLoad = defaultGridOptions.proxyConfig?.autoLoad;
   const enableProxyConfig = options.value.proxyConfig?.enabled;
   if (enableProxyConfig && autoLoad) {
-    props.api.grid.commitProxy?.('_init', formApi.form?.values ?? {});
+    props.api.grid.commitProxy?.('_init', (await formApi.getValues()) ?? {});
     // props.api.reload(formApi.form?.values ?? {});
   }
